+++ src/nav/index.js
@@ -44,14 +44,21 @@ function createNavArray(host, defaults) {
         after(name, config) {
             insert(config, 'after', name);
         },
+        replace(name, config) {
+            const position = items.findIndex(item => item.name === name);
+
+            if (position !== -1) {
+                items[position] = config;
+            } else {
+                insert(config);
+            }
+        },
         remove(name) {
             const position = items.findIndex(item => item.name === name);
 
             if (position !== -1) {
                 return items.splice(position, 1)[0];
             }
-
-            return null;
         }
     });
 }