+++ src/lib.js
@@ -2,6 +2,7 @@ import App from './app/index.js';
 import Widget from './widget/index.js';
 import * as views from './views/index.js';
 import * as pages from './pages/index.js';
+import inspector from './inspector/index.js';
 import router from './core/router.js';
 import utils from './core/utils/index.js';
 
@@ -10,6 +11,7 @@ export {
     App,
     views,
     pages,
+    inspector,
     router,
     utils
 };